+++ app/controller/App.js
@@ -1,23 +1,15 @@
+import Carousel from "./carousel";
+// import CarouselView from "../view/carouselView";
 
-import Carousel from './Carousel';
-import FormModel from '../model/formmodel';
 
 export default class App {
     constructor()
     {
         console.log("Creating the App");
-        // this.CarouselCellView1 = new CarouselCellView1();
-        // this.CarouselCellView2 = new CarouselCellView2();
-        // this.CarouselCellView3 = new CarouselCellView3();
-        // this.CarouselCellView4 = new CarouselCellView4();
-        // this.CarouselView = new CarouselView(this.CarouselCellView1, this.CarouselCellView2, this.CarouselCellView3, this.CarouselCellView4);
-        // this.Carousel = new Carousel(this);
-        // this.FormModel = new FormModel();
+        this.Carousel = new Carousel();
     }
 
     init(){
-
       console.log("in app init");
-      this.Carousel.createCarousel();
     }
 }
