+++ app/index.js
@@ -2,42 +2,17 @@
 import Slider from './js/slider-ui.js';
 import App from './controller/App';
 
+require('./main.scss');
+
 let indexer = require('file-loader?name=../dist/index.html!./index.html');
-// let app = new App();
-let utils = window.fizzyUIUtils;
-let carousel = document.querySelector('.main-carousel');
-let flkty = new Flickity( carousel, {
-  prevNextButtons: false,
-  contain: true,
-  pageDots: false,
-  draggable: false
-});
-let nextButton = document.querySelector('.button--next');
-nextButton.addEventListener( 'click', function(event) {
-  event.preventDefault();
-  flkty.next();
-});
-let previousButton1 = document.querySelector('.back1');
-previousButton1.addEventListener( 'click', function(event) {
-  event.preventDefault();
-  flkty.previous();
-});
-let getReportButton = document.querySelector('.button--report');
-getReportButton.addEventListener( 'click', function(event) {
-  event.preventDefault();
-  flkty.next();
-});
-let previousButton2 = document.querySelector('.back2');
-previousButton2.addEventListener( 'click', function(event) {
-  event.preventDefault();
-  flkty.previous();
-});
-let doneButton = document.querySelector('.button--done');
-doneButton.addEventListener( 'click', function(event) {
-  event.preventDefault();
-  flkty.next();
-});
-// app.init();
+
+
+let app = new App();
+app.init();
 let slider = new Slider();
 
-require('./main.scss');
+
+$( document ).ready(function() {
+  console.log("on ready");
+
+});
